+++ packages/DouYuRecorder/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+- 触发标题黑名单设定额外状态
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`
 
 # 1.7.1