+++ packages/DouYuRecorder/CHANGELOG.md
@@ -3,6 +3,7 @@
 - 触发标题黑名单设定额外状态
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`
 - 尽可能避免scdn
+- 录播姬引擎支持
 
 # 1.7.1