+++ packages/DouYuRecorder/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 1.8.0
 
 - 触发标题黑名单设定额外状态
 - 新增`debugLevel`参数，支持`none`、`basic`、`verbose`